Перенос остатков раздела "Преподаватели" в отдельный модуль

+++ core/data/src/main/kotlin/ru/vo1dmain/timetables/data/entities/instructor/InstructorsRepository.kt
@@ -18,7 +18,7 @@ class InstructorsRepository(application: Application) :
         return dao.findForSubject(subjectId)
     }
     
-    suspend fun deleteByIds(ids: List<Long>) {
+    suspend fun deleteByIds(ids: List<Int>) {
         dao.deleteByIds(ids)
     }
